requirements.txt
flask
inky
Pillow
qrcode
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests